Set title/body from commit-message if given
If `title`, `body`, and `body-path` are all not given, and
`commit-message` is, parse the `title` and `body` from the
'commit-message'.

A properly formatted message (title, blank line, then body) will be used
as title and body. Otherwise, the first line is used as title and the
remaining lines as body. Another valid approach could be to take all the
lines as title, or all the lines until the first blank as title, but the
former is unlikely to be the user's intention and the latter would
introduce complexity.

The parsing function was added in `utils` to facilitate testing.